// src/app/Models/CartDetails.ts
export interface CartDetails {
    cartId: number;
    productId: number;
    quantity: number;
    size: string;
    name: string;
    description: string;
    price: number;
    discount: number;
    stockStatus: string;
    rating: number;
    category: string;
    brand: string;
    imageUrl: string;
    createdAt: string;
    userId: number;
  }